refactor: products-managerを3ファイルに分割
- product-form.tsx: 商品作成/編集フォーム（フォームstate内部管理）
- product-card.tsx: 商品カード + バリエーション管理 + 削除確認ダイアログ
- products-manager.tsx: オーケストレーター（productsリスト管理のみ）
- 760行のモノリシックコンポーネントを責務ごとに分離
